+++ helper.c
@@ -10,8 +10,8 @@ void HELPER(prepare_block_for_execution)(void* tb)
 {
   cpu->current_tb = (TranslationBlock*)tb;
 
-  uint32_t instructions_left = cpu->instructions_count_threshold - cpu->instructions_count_value;
-  uint32_t current_block_size = cpu->current_tb->icount;
+  uint64_t instructions_left = cpu->instructions_count_threshold - cpu->instructions_count_value;
+  uint64_t current_block_size = cpu->current_tb->icount;
 
   if(instructions_left == 0)
   {
